requirements.txt
appdirs==1.4.4
autopep8==1.5.7
black==21.6b0
click==8.0.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
filelock==3.0.12
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
pbr==5.6.0
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whitenoise==5.2.0
